change get_map_num_led() to support bigger than 256 LEDs

+++ reveil/mappage_led.h
@@ -2,9 +2,8 @@
 
 struct pixel_position
 {
-  int x;
-  int y;
-  
+  uint16_t x;
+  uint16_t y;
 };
 
 struct pixel_color
